Use sub-pixel precision during glyph layout

+++ src/ft2_interface.h
@@ -11,7 +11,7 @@ FT_Error debugPrintTest();
 FT_Error renderGlyph(
     FT_Face face,
     FT_ULong char_code,
-    unsigned int *advance_horz_out,
+    float *advance_horz_out,
     int *bearing_x_out,
     int *bearing_y_out,
     unsigned int *bitmap_width_out,
